+++ sdk/appconfiguration/app-configuration/samples/getSettingOnlyIfChanged.ts
@@ -45,7 +45,7 @@ export async function run() {
 
 async function cleanupSampleValues(keys: string[], client: AppConfigurationClient) {
   const existingSettings = await client.listConfigurationSettings({
-    keys: keys
+    keyFilter: keys.join(',')
   });
 
   for await (const setting of existingSettings) {